# Documentation/apiDoc.yaml
swagger: "2.0"
info:
  version: 1.0.0
  title: Shared Agenda Application
host: sharedagendaapp.herokuapp.com
schemes:
  - https
consumes:
  - application/json
produces:
  - application/json
paths:
...
  updateEvent:
    post:
      operationId: updateEvent
      tags:
        - Events
      requestBody:
        content:
          application/json:
            schema:
              $ref: '../Backend/schema/src/yaml/requests/UpdateEventRequest.yaml'
      responses:
        "200":
          description: OK
          schema:
            $ref: '../Backend/schema/src/yaml/responses/UpdateEventResponse.yaml'
        "400":
          description: "Invalid Parameter: eventId, eventname, starterId, and type cannot be null or empty"
        "404":
          description: "Not Found: target EventId not found in database"
...
  AddGroupMember:
    post:
      operationId: AddGroupMember
      tags:
        - Groups
      requestBody:
        content:
          application/json:
            schema:
              $ref: '../Backend/schema/src/yaml/requests/AddGroupMemberRequest.yaml'
      responses:
        "200":
          description: Ok
          schema:
            $ref: '../Backend/schema/src/yaml/responses/AddGroupMemberResponse.yaml'
        "400":
          description: "Invalid Parameter: groupid and memberids cannot be null or empty"
        "404":
          description: "Not Found: memberids or groupid do not exist"